task-2/worker/requirements.txt
boto3==1.39.13
psycopg2-binary==2.9.9